requirements.txt
docopt==0.6.2
numpy==1.17.4
Pillow==6.2.1